put the personality count in an enum instead
